rabbit MQ integrated in company app -> IDEA is when any person add review of the company by the help of message queue an Asynchoronus communication is established and an Average rating is updated in the company

+++ MicroServices/Jobs App MicroServiice/Company-MicroService/src/main/java/org/CompanyMicroService/DTOs/review_DTo.java
@@ -1,5 +1,6 @@
 package org.CompanyMicroService.DTOs;
 
+import ReviewsMS.pojo.reviews_pojo;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 import org.bson.types.ObjectId;
@@ -9,6 +10,6 @@ import org.bson.types.ObjectId;
 public class review_DTo {
     ObjectId id;
     String review;
-    int rating;
+    Double rating;
     String companyName;
 }
